backend/pyproject.toml
[tool.poetry]
name = "backend-online-chat"
version = "0.1.0"
description = ""
authors = ["ReaToi <beksultannurbekov0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.128.0"
sqlalchemy = "^2.0.45"
asyncpg = "^0.31.0"
alembic = "^1.17.2"
pydantic-settings = "^2.12.0"
python-jose = "^3.5.0"
greenlet = "^3.3.0"
websockets = "^15.0.1"
uvicorn = "^0.40.0"
authx = "^1.4.3"
passlib = "^1.7.4"
bcrypt = "4.0.1"
pydantic = {extras = ["email"], version = "^2.12.5"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
